requirements/requirements.txt
Django==3.0.6
djangorestframework==3.11.0
psycopg2==2.8.5
dj-database-url==0.5.0